% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDataTableWorksheet.R
\name{addDataTableWorksheet}
\alias{addDataTableWorksheet}
\title{Add a new worksheet to the excel workbook and export the given data table x to it nicely formatted.}
\usage{
addDataTableWorksheet(
  wb,
  sheet,
  title = NULL,
  x = NULL,
  loopdim = 3,
  style = NULL,
  tableStyle = "TableStyleMedium17",
  comparisons = NULL,
  comparisonStyle = NULL,
  ...
)
}
\description{
Add a new worksheet to the excel workbook and export the given data table x to it nicely formatted.
}
