% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{age.exactRounded}
\alias{age.exactRounded}
\title{Calculate the age of the insured based on exact age at contract closing, rounded
to the nearest birthday.}
\usage{
age.exactRounded(params, values)
}
\arguments{
\item{params}{The parameters of the contract.}

\item{values}{Unused by default (already calculated values of the contract)}
}
\description{
Calculate the age of the insured based on exact age at contract closing, rounded
to the nearest birthday.
}
