% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{age.yearDifference}
\alias{age.yearDifference}
\title{Calculate the age of the insured based on the difference of the bith year and
contract closing year.}
\usage{
age.yearDifference(params, values)
}
\arguments{
\item{params}{The parameters of the contract.}

\item{values}{Unused by default (already calculated values of the contract)}
}
\description{
Calculate the age of the insured based on the difference of the bith year and
contract closing year.
}
