% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\name{costs.baseAlpha}
\alias{costs.baseAlpha}
\title{Helper function to define base costs with base alpha, but otherwise unchanged costs}
\usage{
costs.baseAlpha(alpha)
}
\arguments{
\item{alpha}{The minimum alpha / Zillmer cost that cannot be waived}
}
\description{
Returns a function that sets base alpha (and Zillmer) costs to the given value,
but otherwise uses the full costs defined by the Costs parameter.
}
\details{
This function can be set as minCosts parameter for a tariff and makes sure
that only alpha costs are modified / waived, but no other costs.
}
