% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInsuranceContract_xlsx.R
\name{costValuesAsDF}
\alias{costValuesAsDF}
\title{Convert the multi-dimensional costs array to a data.frame for output to a file}
\usage{
costValuesAsDF(costValues)
}
\arguments{
\item{costValues}{Cost definition data structure}
}
\description{
Convert the cost values array to a tx15 matrix
}
\details{
Not to be called directly, but implicitly by the \link{InsuranceContract} object.
Convert the array containing cost values like cashflows, present
values, etc. (objects of dimension tx5x3) to a matrix with dimensions (tx15)
}
