% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{fallbackFields}
\alias{fallbackFields}
\title{fallbackFields(fields, fallback)}
\usage{
fallbackFields(fields, valuelist)
}
\description{
Replace all missing values in fields (either missing or NA) with
their corresponding values from fallback. Members in fallback that are missing
in fields are inserted
}

