% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{fillNAgaps}
\alias{fillNAgaps}
\title{Taken from the R Cookbook:
http://www.cookbook-r.com/Manipulating_data/Filling_in_NAs_with_last_non-NA_value/
LICENSE (from that page): The R code is freely available for use without any restrictions.
In other words: you may reuse the R code for any purpose (and under any license).}
\usage{
fillNAgaps(x, firstBack = FALSE)
}
\description{
Taken from the R Cookbook:
http://www.cookbook-r.com/Manipulating_data/Filling_in_NAs_with_last_non-NA_value/
LICENSE (from that page): The R code is freely available for use without any restrictions.
In other words: you may reuse the R code for any purpose (and under any license).
}
