% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfitParticipation.R
\name{filterProfitRates}
\alias{filterProfitRates}
\title{Filter the whole data.frame of profit rates for the given profit classes}
\usage{
filterProfitRates(rates, classes)
}
\arguments{
\item{rates}{data.frame containing all profit rates for multiple profit classes}

\item{classes}{the profit classes, for which rates should be extracted}
}
\description{
This is a rather trivial helper function, which just calls \code{\link[dplyr:filter]{dplyr::filter()}}.
}
