% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegisterLITariff.R
\name{get.tariff}
\alias{get.tariff}
\title{Retrieve a LifeInsureR tariff registered with the package}
\usage{
get.tariff(Company, Produkt, GV = NULL, AV = NULL)
}
\arguments{
\item{Company}{The (short) identifier for the company. Must contain only letters, digits and periods (i.e. a valid R identifier, given as string).}

\item{Produkt}{The short product/tariff ID to identify the tariff.}

\item{GV}{The (optional) profit group ("Gewinnverband") for the product.}

\item{AV}{The (optional) profit sub-group ("Abrechnungsveband") for the product}
}
\value{
The \link{InsuranceTarif} object defining the tariff/product.
}
\description{
While the package allows individual tariff object to be created and used,
one can also register the tariffs for each company with the package.
}
\details{
After registering the tariff/product with the function \code{\link[=register.tariff]{register.tariff()}},
this function can be used to retrieve the \link{InsuranceTarif} object using
the given ID.
}
\examples{
library(MortalityTables)
library(magrittr)
mortalityTables.load("Austria_Annuities_AVOe2005R")
InsuranceTarif$new(
    type = "annuity",
    # tarif = "Ann1",
    # desc = "An annuity with single-premium",
    premiumPeriod = 1,
    mortalityTable = AVOe2005R.unisex,
    i = 0.005
) \%>\%
     register.tariff(Company = "ExVAG", Produkt = "ImmAnn1") \%>\%
     register.tariff(Company = "ExVAG", Produkt = "ImmAnn1_BV")

 Tariff.ImmAnnuity = get.tariff("ExVAG", "ImmAnn1")
}
