% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\name{initializeCosts}
\alias{initializeCosts}
\title{Initialize a cost matrix with dimensions: [CostType, Basis, Period], with:
    CostType: alpha, Zillmer, beta, gamma, gamma_nopremiums, unitcosts
    Basis:    SumInsured, SumPremiums, GrossPremium, NetPremium, Constant
    Period:   once, PremiumPeriod, PremiumFree, PolicyPeriod
TODO: gamma an Erlebensleistungen?}
\usage{
initializeCosts(costs, alpha, Zillmer, beta, gamma, gamma.paidUp,
  gamma.premiumfree, unitcosts, unitcosts.PolicyPeriod)
}
\description{
Initialize a cost matrix with dimensions: [CostType, Basis, Period], with:
    CostType: alpha, Zillmer, beta, gamma, gamma_nopremiums, unitcosts
    Basis:    SumInsured, SumPremiums, GrossPremium, NetPremium, Constant
    Period:   once, PremiumPeriod, PremiumFree, PolicyPeriod
TODO: gamma an Erlebensleistungen?
}
