% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\docType{data}
\name{InsuranceContract.ParameterDefaults}
\alias{InsuranceContract.ParameterDefaults}
\title{Default parameters for the InsuranceContract class. A new contract will be
pre-filled with these values, and values passed in the constructor (or with
other setter functions) will override these values.}
\format{An object of class \code{list} of length 7.}
\usage{
InsuranceContract.ParameterDefaults
}
\description{
Default parameters for the InsuranceContract class. A new contract will be
pre-filled with these values, and values passed in the constructor (or with
other setter functions) will override these values.
}
\keyword{datasets}
