% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\name{InsuranceContract.ParametersFallback}
\alias{InsuranceContract.ParametersFallback}
\title{Use fallback values for missing contract parameters}
\usage{
InsuranceContract.ParametersFallback(params, fallback, ppParameters = TRUE)
}
\arguments{
\item{params}{Current, explicitly set contract parameters. All NULL values
will be filled with the corresponding entry from \code{fallback}.}

\item{fallback}{Fallback values that will be used when the corresponding
entry in \code{params} is NULL.}

\item{ppParameters}{Whether profit participation parameters should also be
filled (default is TRUE)}
}
\description{
Provide default values for the insurance contract parameters if any of the
parameters is not explicitly set.
}
