% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\name{InsuranceContract.ParametersFill}
\alias{InsuranceContract.ParametersFill}
\title{InsuranceContract.ParametersFill}
\usage{
InsuranceContract.ParametersFill(
  params = InsuranceContract.ParameterStructure,
  costs = NULL,
  ...
)
}
\arguments{
\item{params}{Initial values of the insurance contract parameters. (default: empty parameter structure)}

\item{costs, ...}{Values for any of the entries in the insurance contract
parameter structure. These values take precedence over the
initial parameters provided in \code{params}.}
}
\description{
Initialize the insurance contract parameters from the passed
arguments. Arguments not given are left unchanged. If no existing parameter
structure is given, an empty (i.e. all NULL entries) structure is used.
}
