% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\docType{data}
\name{InsuranceContract.ParameterStructure}
\alias{InsuranceContract.ParameterStructure}
\title{Full insurance contract parameter structure.}
\format{
An object of class \code{list} of length 8.
}
\usage{
InsuranceContract.ParameterStructure
}
\description{
All values are filled with NULL,
so the functions \code{\link{InsuranceContract.ParametersFill}} and
\code{\link{InsuranceContract.ParametersFallback}} can be used to override
existing parameters or to provide default values for unset (NULL) entries.
}
\keyword{datasets}
