% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceContract.R
\name{InsuranceContract}
\alias{InsuranceContract}
\title{Base Class for Insurance Contracts}
\description{
R6 class that models a complete, general insurance contract.
The corresponding tariff and the profit participation scheme, as well as
all other relevant contract parameters (if not defined by the tariff or
explicitly overridden by the contract) can be given in the constructor.
}
\details{
Immediately upon construction, all premiums, reserves and cash flows for the
whole contract period are calculated.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InsuranceContract$new()}}
\item \href{#method-addHistorySnapshot}{\code{InsuranceContract$addHistorySnapshot()}}
\item \href{#method-addBlock}{\code{InsuranceContract$addBlock()}}
\item \href{#method-addDynamics}{\code{InsuranceContract$addDynamics()}}
\item \href{#method-calculateContract}{\code{InsuranceContract$calculateContract()}}
\item \href{#method-consolidateBlocks}{\code{InsuranceContract$consolidateBlocks()}}
\item \href{#method-premiumWaiver}{\code{InsuranceContract$premiumWaiver()}}
\item \href{#method-premiumWaiverOLD}{\code{InsuranceContract$premiumWaiverOLD()}}
\item \href{#method-profitScenario}{\code{InsuranceContract$profitScenario()}}
\item \href{#method-addProfitScenario}{\code{InsuranceContract$addProfitScenario()}}
\item \href{#method-clone}{\code{InsuranceContract$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$new(
  tarif,
  parent = NULL,
  calculate = "all",
  profitid = "default",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addHistorySnapshot"></a>}}
\if{latex}{\out{\hypertarget{method-addHistorySnapshot}{}}}
\subsection{Method \code{addHistorySnapshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$addHistorySnapshot(
  time = 0,
  comment = "Initial contract values",
  type = "Contract",
  params = self$Parameters,
  values = self$Values,
  tarif = self$tarif,
  blocks = self$blocks
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addBlock"></a>}}
\if{latex}{\out{\hypertarget{method-addBlock}{}}}
\subsection{Method \code{addBlock()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$addBlock(
  id = NULL,
  block = NULL,
  t = block$Values$int$blockStart,
  comment = comment,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addDynamics"></a>}}
\if{latex}{\out{\hypertarget{method-addDynamics}{}}}
\subsection{Method \code{addDynamics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$addDynamics(t, NewSumInsured, SumInsuredDelta, id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateContract"></a>}}
\if{latex}{\out{\hypertarget{method-calculateContract}{}}}
\subsection{Method \code{calculateContract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$calculateContract(
  calculate = "all",
  valuesFrom = 0,
  premiumCalculationTime = 0,
  preservePastPV = TRUE,
  recalculatePremiums = TRUE,
  recalculatePremiumSum = TRUE,
  history_comment = NULL,
  history_type = "Contract"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidateBlocks"></a>}}
\if{latex}{\out{\hypertarget{method-consolidateBlocks}{}}}
\subsection{Method \code{consolidateBlocks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$consolidateBlocks(valuesFrom = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-premiumWaiver"></a>}}
\if{latex}{\out{\hypertarget{method-premiumWaiver}{}}}
\subsection{Method \code{premiumWaiver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$premiumWaiver(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-premiumWaiverOLD"></a>}}
\if{latex}{\out{\hypertarget{method-premiumWaiverOLD}{}}}
\subsection{Method \code{premiumWaiverOLD()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$premiumWaiverOLD(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-profitScenario"></a>}}
\if{latex}{\out{\hypertarget{method-profitScenario}{}}}
\subsection{Method \code{profitScenario()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$profitScenario(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addProfitScenario"></a>}}
\if{latex}{\out{\hypertarget{method-addProfitScenario}{}}}
\subsection{Method \code{addProfitScenario()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$addProfitScenario(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceContract$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
