% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\docType{data}
\name{InsuranceContract.Values}
\alias{InsuranceContract.Values}
\title{Data structure (filled only with NULL) for insurance contract class member values.}
\format{
An object of class \code{list} of length 16.
}
\usage{
InsuranceContract.Values
}
\description{
Data structure (filled only with NULL) for insurance contract class member values.
}
\keyword{datasets}
