% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceTarif.R
\docType{data}
\name{InsuranceTarif}
\alias{InsuranceTarif}
\title{Base class for Insurance Tarifs, providing calculation functions to the contract}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InsuranceTarif
}
\description{
This is a base class for holding contract-independent values and
providing methods to calculate cash flows, premiums, etc. Objects of this
class do NOT contain contract-specific values like age, death probabilities,
premiums, reserves, etc. Rather, they are the calculation kernels that will
be called by the \code{\link{InsuranceContract}} objects to make the actual,
tariff-specific calculations.
}
\keyword{datasets}

