% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{isRegularPremiumContract}
\alias{isRegularPremiumContract}
\title{Determine whether a contract (given all parameters) is a contract with regular premiums}
\usage{
isRegularPremiumContract(params, values)
}
\arguments{
\item{params}{The parameters of the contract.}

\item{values}{Unused by default (already calculated values of the contract)}
}
\description{
Regular premium contracts are identified by the parameter \code{premiumPeriod > 1}.
}
