% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{isSinglePremiumContract}
\alias{isSinglePremiumContract}
\title{Determine whether a contract (given all parameters) is a single-premium contract or with regular premiums}
\usage{
isSinglePremiumContract(params, values)
}
\arguments{
\item{params}{The parameters of the contract.}

\item{values}{Unused by default (already calculated values of the contract)}
}
\description{
The death benefit will be the full sumInsured for the first year after the
deferral period and then decrease linearly to 0 at the end of the policyPeriod.
This can be used with the \code{deathBenefit} parameter for insurance
contracts, but should not be called directly.
}
