% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractGrid.R
\name{makeContractGridDimname.InsuranceTarif}
\alias{makeContractGridDimname.InsuranceTarif}
\alias{makeContractGridDimname.R6}
\alias{makeContractGridDimname.mortalityTable}
\alias{makeContractGridDimname.numeric}
\alias{makeContractGridDimname.double}
\alias{makeContractGridDimname.default}
\alias{makeContractGridDimname}
\title{Create human-readable labels for the dimensions in a \code{\link[=contractGrid]{contractGrid()}}}
\usage{
\method{makeContractGridDimname}{InsuranceTarif}(value)

\method{makeContractGridDimname}{R6}(value)

\method{makeContractGridDimname}{mortalityTable}(value)

\method{makeContractGridDimname}{numeric}(value)

\method{makeContractGridDimname}{double}(value)

\method{makeContractGridDimname}{default}(value)

makeContractGridDimname(value)
}
\arguments{
\item{value}{the value along the axis, for which a name should be generated}
}
\description{
The funciton \code{makeContractGridDimname} generates a short, human-readable
dimension label for the object passed as \code{value}, to be used in \link{contractGrid}.
The default is to use the \code{value} unchanged as the row-/columnname, but
for some parameter values (like a \link{InsuranceTarif} or \link{MortalityTable})
a custom method of this function is needed to create the (short) human-readable
representation for the axes in the grid.
}
\section{Methods (by class)}{
\itemize{
\item \code{InsuranceTarif}: Create a dimensional name for an \link{InsuranceTarif} object

\item \code{R6}: Create a dimensional name for an R6 object (using its \code{name} field)

\item \code{mortalityTable}: Create a dimensional name for an \link{MortalityTable} object

\item \code{numeric}: Create a dimensional name for a numeric parameter value

\item \code{double}: Create a dimensional name for a numeric parameter value

\item \code{default}: Create a dimensional name for an object that can be directly used as a human-readable row-/columnname
}}

