% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{padLast}
\alias{padLast}
\title{Pad the vector \code{v} to length \code{l} by repeating the last entry of the
vector.}
\usage{
padLast(v, ...)
}
\arguments{
\item{v}{the vector to pad by repeating the last element}

\item{...}{arguments passed through to \code{pad0}}
}
\value{
vector padded to the correct length
}
\description{
This function is just a trivial wrapper around \code{pad0} and only calls \code{\link[=pad0]{pad0()}}
with the last element of the vector as padding value instead of the default 0.
}
\examples{
padLast(1:5, 7) # 5 is repeated twice
padLast(1:5, 3) # no padding needed

}
