% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\docType{class}
\name{PaymentTimeSingleEnum-class}
\alias{PaymentTimeSingleEnum-class}
\alias{PaymentTimeEnum}
\title{Enum to describe when a benefit or premium payment is due (in advance or in arrears)}
\description{
Enum to describe when a benefit or premium payment is due (in advance or in arrears)
}
\details{
Currently, only two values are allowed;
\itemize{
\item "in advance"
\item "in arrears"
}
}
