% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\docType{class}
\name{ProfitComponentsMultipleEnum-class}
\alias{ProfitComponentsMultipleEnum-class}
\alias{ProfitComponentsEnum}
\title{Enum to define the different components of profit participation.}
\description{
Enum to define the different components of profit participation.
}
\details{
Profit participation schemes typically consist of different components,
which are calculated independently. Typical components are interest profit
to distribute investment gains to the customer, risk profit and expense profit
to return security margins in the biometric risk and the expenses to the customer
and sum profit, which aplies to contracts with higher sums insured, where
charged expenses are calculated from the sum insured, while the actual
expenses are more or less constant. Thus, high contracts are charged more,
which causes profits that are returned as sum profit.

As a special case, part of the profits can be stored in a terminal bonus
reserve and only distributed on maturity (or potentially on death), but
not on surrender. Some (older) profit participation schemes add an independently
calculated bonus on maturity (e.g. twice the total profit assignment of the
last year) at maturity to give customers an additional incentive not to
surrender a contract.

Possible values are (multiple can be given):
\itemize{
\item "interest"
\item "risk"
\item "expense"
\item "sum"
\item "terminal"
\item "TBF"
}
}
