% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfitParticipation.R
\name{ProfitParticipation}
\alias{ProfitParticipation}
\title{Base Class for Profit Participation Schemes}
\description{
base class for Profit Participation schemes  (holding contract-independent values and
providing methods to calculate the profit participation values from the given
reserves).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProfitParticipation$new()}}
\item \href{#method-setParameters}{\code{ProfitParticipation$setParameters()}}
\item \href{#method-setFunctions}{\code{ProfitParticipation$setFunctions()}}
\item \href{#method-setFallbackParameters}{\code{ProfitParticipation$setFallbackParameters()}}
\item \href{#method-createModification}{\code{ProfitParticipation$createModification()}}
\item \href{#method-getAdvanceProfitParticipation}{\code{ProfitParticipation$getAdvanceProfitParticipation()}}
\item \href{#method-getAdvanceProfitParticipationAfterUnitCosts}{\code{ProfitParticipation$getAdvanceProfitParticipationAfterUnitCosts()}}
\item \href{#method-setupRates}{\code{ProfitParticipation$setupRates()}}
\item \href{#method-adjustRates}{\code{ProfitParticipation$adjustRates()}}
\item \href{#method-getProfitParticipation}{\code{ProfitParticipation$getProfitParticipation()}}
\item \href{#method-clone}{\code{ProfitParticipation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$new(name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setParameters"></a>}}
\if{latex}{\out{\hypertarget{method-setParameters}{}}}
\subsection{Method \code{setParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$setParameters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setFunctions"></a>}}
\if{latex}{\out{\hypertarget{method-setFunctions}{}}}
\subsection{Method \code{setFunctions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$setFunctions(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setFallbackParameters"></a>}}
\if{latex}{\out{\hypertarget{method-setFallbackParameters}{}}}
\subsection{Method \code{setFallbackParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$setFallbackParameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createModification"></a>}}
\if{latex}{\out{\hypertarget{method-createModification}{}}}
\subsection{Method \code{createModification()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$createModification(name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAdvanceProfitParticipation"></a>}}
\if{latex}{\out{\hypertarget{method-getAdvanceProfitParticipation}{}}}
\subsection{Method \code{getAdvanceProfitParticipation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$getAdvanceProfitParticipation(params, values, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAdvanceProfitParticipationAfterUnitCosts"></a>}}
\if{latex}{\out{\hypertarget{method-getAdvanceProfitParticipationAfterUnitCosts}{}}}
\subsection{Method \code{getAdvanceProfitParticipationAfterUnitCosts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$getAdvanceProfitParticipationAfterUnitCosts(
  params,
  values,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupRates"></a>}}
\if{latex}{\out{\hypertarget{method-setupRates}{}}}
\subsection{Method \code{setupRates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$setupRates(params, values, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adjustRates"></a>}}
\if{latex}{\out{\hypertarget{method-adjustRates}{}}}
\subsection{Method \code{adjustRates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$adjustRates(rates, params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProfitParticipation"></a>}}
\if{latex}{\out{\hypertarget{method-getProfitParticipation}{}}}
\subsection{Method \code{getProfitParticipation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$getProfitParticipation(
  calculateFrom = 0,
  profitScenario = NULL,
  params,
  values,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfitParticipation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
