% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfitParticipation_Functions.R
\name{ProfitParticipationFunctions}
\alias{ProfitParticipationFunctions}
\alias{profPart.base.PreviousZillmerReserve}
\alias{profPart.base.ZillmerReserveT2}
\alias{profPart.base.ZillmerRiskPremium}
\alias{profPart.base.contractualReserve}
\alias{profPart.base.meanContractualReserve}
\alias{profPart.base.sumInsured}
\alias{profPart.benefit.Profit}
\alias{profPart.benefit.ProfitPlusGuaranteedInterest}
\alias{profPart.benefit.ProfitPlusHalfGuaranteedInterest}
\alias{profPart.benefit.ProfitPlusHalfInterestMinGuaranteeTotal}
\alias{profPart.benefit.ProfitPlusHalfTotalInterest}
\alias{profPart.benefit.ProfitPlusInterestMinGuaranteeTotal}
\alias{profPart.benefit.ProfitPlusTerminalBonusReserve}
\alias{profPart.benefit.ProfitPlusTotalInterest}
\alias{profPart.benefit.TerminalBonus5YearsProRata}
\alias{profPart.calculate.RateOnBase}
\alias{profPart.calculate.RateOnBaseSGFFactor}
\alias{profPart.calculate.RatePlusGuaranteeOnBase}
\alias{profPart.rate.expenseProfit}
\alias{profPart.rate.interestProfit}
\alias{profPart.rate.interestProfit2}
\alias{profPart.rate.interestProfitPlusGuarantee}
\alias{profPart.rate.riskProfit}
\alias{profPart.rate.sumProfit}
\alias{profPart.rate.terminalBonus}
\alias{profPart.rate.terminalBonusFundRatio}
\alias{profPart.rate.totalInterest}
\alias{profPart.rate.totalInterest2}
\title{Helper functions for profit participation}
\usage{
profPart.base.PreviousZillmerReserve(rates, params, values, ...)

profPart.base.ZillmerReserveT2(rates, params, values, ...)

profPart.base.contractualReserve(rates, params, values, ...)

profPart.base.meanContractualReserve(rates, params, values, ...)

profPart.base.ZillmerRiskPremium(rates, params, values, ...)

profPart.base.sumInsured(rates, params, values, ...)

profPart.rate.interestProfit(rates, params, values, ...)

profPart.rate.riskProfit(rates, params, values, ...)

profPart.rate.expenseProfit(rates, params, values, ...)

profPart.rate.sumProfit(rates, params, values, ...)

profPart.rate.terminalBonus(rates, params, values, ...)

profPart.rate.terminalBonusFundRatio(rates, params, values, ...)

profPart.rate.interestProfitPlusGuarantee(rates, params, values, ...)

profPart.rate.totalInterest(rates, params, values, ...)

profPart.rate.totalInterest2(rates, params, values, ...)

profPart.rate.interestProfit2(rates, params, values, ...)

profPart.calculate.RateOnBase(base, rate, waiting, rates, params, values, ...)

profPart.calculate.RatePlusGuaranteeOnBase(base, rate, waiting, rates, params,
  values, ...)

profPart.calculate.RateOnBaseSGFFactor(base, rate, waiting, rates, params,
  values, ...)

profPart.benefit.ProfitPlusTerminalBonusReserve(profits, rates, params, values)

profPart.benefit.Profit(profits, rates, params, values)

profPart.benefit.ProfitPlusGuaranteedInterest(profits, rates, params, values)

profPart.benefit.ProfitPlusTotalInterest(profits, rates, params, values)

profPart.benefit.ProfitPlusHalfTotalInterest(profits, rates, params, values)

profPart.benefit.ProfitPlusHalfGuaranteedInterest(profits, rates, params,
  values)

profPart.benefit.ProfitPlusInterestMinGuaranteeTotal(profits, rates, params,
  values)

profPart.benefit.ProfitPlusHalfInterestMinGuaranteeTotal(profits, rates, params,
  values)

profPart.benefit.TerminalBonus5YearsProRata(profits, rates, params, values)
}
\description{
Various helper functions for the \code{ProfitParticipation} class that
provide the building blocks for the individual components of profit participation,
the rates and how the assigned profit is calculated.
}
\section{Functions}{
\itemize{
\item \code{profPart.base.PreviousZillmerReserve}: Basis for profit: Previous Zillmer reserve (no administration cost reserve)

\item \code{profPart.base.ZillmerReserveT2}: Basis for profit: Zillmer reserve (no administration cost reserve) at time t-2

\item \code{profPart.base.contractualReserve}: Basis for profit: Contractual reserve (including administration costs) at time t

\item \code{profPart.base.meanContractualReserve}: Basis for profit: Contractual reserve (including administration costs) averaged over t and t-1

\item \code{profPart.base.ZillmerRiskPremium}: Basis for risk/mortality profit: Zillmer Risk Premium of the past year

\item \code{profPart.base.sumInsured}: Basis for expense/sum profit: sum insured

\item \code{profPart.rate.interestProfit}: Returns the array of interest profit rates (keyed by year)

\item \code{profPart.rate.riskProfit}: Returns the array of risk profit rates (keyed by year)

\item \code{profPart.rate.expenseProfit}: Returns the array of expense profit rates (keyed by year)

\item \code{profPart.rate.sumProfit}: Returns the array of sum profit rates (keyed by year)

\item \code{profPart.rate.terminalBonus}: Returns the array of terminal bonus rates (keyed by year)

\item \code{profPart.rate.terminalBonusFundRatio}: Returns the array of terminal bonus rates (keyed by year) as the terminal bonus fund ratio

\item \code{profPart.rate.interestProfitPlusGuarantee}: Rate for interest on past profits: total interest rate

\item \code{profPart.rate.totalInterest}: Rate for interest on past profits: total interest rate

\item \code{profPart.rate.totalInterest2}: Rate for interest on past profits: second total interest rate

\item \code{profPart.rate.interestProfit2}: Rate for interest on past profits: second interest profit rate (not including guaranteed interest), keyed by year

\item \code{profPart.calculate.RateOnBase}: Calculate profit by a simple rate applied on the basis (with an optional waiting vector of values 0 or 1)

\item \code{profPart.calculate.RatePlusGuaranteeOnBase}: Calculate profit by a rate + guaranteed interest applied on the basis (with an optional waiting vector of values 0 or 1)

\item \code{profPart.calculate.RateOnBaseSGFFactor}: Calculate profit by a simple rate applied on the basis (with only (1-SGFFactor) put into profit participation, and an optional waiting vector of values 0 or 1)

\item \code{profPart.benefit.ProfitPlusTerminalBonusReserve}: Calculate survival benefit as total profit amount plus the terminal bonus reserve

\item \code{profPart.benefit.Profit}: Calculate benefit as total profit accrued so far

\item \code{profPart.benefit.ProfitPlusGuaranteedInterest}: Calculate accrued death benefit as total profit with (guaranteed) interest for one year

\item \code{profPart.benefit.ProfitPlusTotalInterest}: Calculate accrued death benefit as total profit with total interest (interest on profit rate) for one year

\item \code{profPart.benefit.ProfitPlusHalfTotalInterest}: Calculate accrued benefit as total profit with total interest (interest on profit rate) for half a year

\item \code{profPart.benefit.ProfitPlusHalfGuaranteedInterest}: Calculate death benefit as total profit with (guaranteed) interest for one year

\item \code{profPart.benefit.ProfitPlusInterestMinGuaranteeTotal}: Calculate accrued benefit as total profit with interest for one year (min of guarantee and total interest)

\item \code{profPart.benefit.ProfitPlusHalfInterestMinGuaranteeTotal}: Calculate accrued benefit as total profit with interest for half a year (min of guarantee and total interest)

\item \code{profPart.benefit.TerminalBonus5YearsProRata}: Calculate benefit from terminal bonus as 1/n parts of the terminal bonus reserve during the last 5 years
}}

