% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegisterLITariff.R
\name{register.company}
\alias{register.company}
\title{Register a company with a given shortcut}
\usage{
register.company(Company, ...)
}
\arguments{
\item{Company}{The (short) identifier for the company. Must contain only letters, digits and periods (i.e. a valid R identifier, given as string).}

\item{...}{Additional parameters to be used for creating tariff templates, etc.
Potential arguments are \code{name} for the full company name, \code{include}
to define the company-specific include file for new product definitions.}
}
\value{
None
}
\description{
While the package allows individual tariff object to be created and used,
one can also register the tariffs for each company with the package.
}
\details{
This function sets up the company data (including the short name) and other
parameters/features used for templating. Once a company is registered with a
short name (and optionally its long name and other parameters), one can register
tariffs/producs with the \code{\link[=register.tariff]{register.tariff()}} function.
}
\examples{
register.company("ExVAG", name = "Example Versicherung AG")

}
