% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegisterLITariff.R
\name{register.tariff}
\alias{register.tariff}
\title{Register a tariff with the package}
\usage{
register.tariff(Tarif, Company, Produkt, GV = NULL, AV = NULL)
}
\arguments{
\item{Tarif}{The \link{LifeInsuranceTarif} object defining the tariff/product.}

\item{Company}{The (short) identifier for the company. Must contain only letters, digits and periods (i.e. a valid R identifier, given as string).}

\item{Produkt}{The short product/tariff ID to identify the tariff.}

\item{GV}{The (optional) profit group ("Gewinnverband") for the product.}

\item{AV}{The (optional) profit sub-group ("Abrechnungsveband") for the product}
}
\value{
The Tarif is returned, so this function can be chained to register the same object with multiple names or GV/AB combinations.
}
\description{
While the package allows individual tariff object to be created and used,
one can also register the tariffs for each company with the package.
}
\details{
This function registers a \link{LifeInsuranceTarif} object with the \code{LiveInsureR}
package using a given ID, so automated scripts can then retrieve the
products/tariffs based on those IDs (e.g. tariff IDs used in the policy
administration systems) using the function \code{\link[=get.tariff]{get.tariff()}}.
}
\examples{
library(MortalityTables)
library(magrittr)
mortalityTables.load("Austria_Annuities_AVOe2005R")
 Tariff.ImmAnnuity = InsuranceTarif$new(
    type = "annuity",
    # tarif = "Ann1",
    # desc = "An annuity with single-premium",
    premiumPeriod = 1,
    mortalityTable = AVOe2005R.unisex,
    i = 0.005
) \%>\%
     register.tariff(Company = "ExVAG", Produkt = "ImmAnn1") \%>\%
     register.tariff(Company = "ExVAG", Produkt = "ImmAnn1_BV")

}
