% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{rollingmean}
\alias{rollingmean}
\title{Calculate the rolling mean of length 2}
\usage{
rollingmean(x)
}
\arguments{
\item{x}{vector of values, for which the rolling mean is calculated}
}
\description{
Calculate the rolling mean of length 2
}
\examples{
rollingmean(1:10)
}
