% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\docType{class}
\name{SexSingleEnum-class}
\alias{SexSingleEnum-class}
\alias{SexEnum}
\title{Enum to describe possble sexes in an insurance contract or tariff.}
\description{
Enum to describe possble sexes in an insurance contract or tariff.
}
\details{
Currently, only possible values are allowed;
\itemize{
\item "unisex"
\item "male"
\item "female"
}
}
