% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceTarif.R
\docType{class}
\name{TariffTypeSingleEnum-class}
\alias{TariffTypeSingleEnum-class}
\alias{TariffTypeEnum}
\title{An enum specifying the main characteristics of the tarif.}
\description{
Possible values are:
\describe{
\item{annuity}{Whole life or term annuity (periodic survival benefits)
with flexible payouts (constand, increasing, decreasing, arbitrary,
etc.)}
\item{wholelife}{A whole or term life insurance with only death benefits.
The benefit can be constant, increasing, decreasing, described by
a function, etc.}
\item{endowment}{An  endowment with death and survival benefits,
potentially with different benefits.}
\item{pureendowment}{A pure endowment with only a survival benefit at
the end of the contract. Optionally, in case of death, all or part
of the premiums paid may be refunded.}
\item{terme-fix}{A terme-fix insurance with a fixed payout at the end
of the contract, even if the insured dies before that time.
Premiums are paid until death of the insured.}
\item{dread-disease}{A dread-disease insurance, which pays in case of
a severe illness (typically heart attacks, cancer, strokes, etc.),
but not in case of death.}
\item{endowment + dread-disease}{A combination of an endowment and a
temporary dread-disease insurance. Benefits occur either on death,
severe illness or survival, whichever comes first.}
}
}
