% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{valueOrFunction}
\alias{valueOrFunction}
\title{If \code{val} is a function, evaluate it, otherwise return \code{val}}
\usage{
valueOrFunction(val, ...)
}
\arguments{
\item{val}{Function or value}

\item{...}{Argument passed to \code{val} if it is a function}
}
\description{
If \code{val} is a function, evaluate it, otherwise return \code{val}
}
\examples{
valueOrFunction(3) # returns 3
valueOrFunction(`+`, 1, 2) # also returns 3
A = `+`
valueOrFunction(A, 1, 2)
}
