% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractGrid.R
\name{contractGrid}
\alias{contractGrid}
\title{Create a multi-dimensional grid of InsuranceContract objects, where the axes
ranges are given with the axes argument.
This function will return the full InsuranceContract objects, so apply can
later be used to extract premiums, reserves and other values to display in
a grid.}
\usage{
contractGrid(
  axes = list(age = seq(20, 60, 10), policyPeriod = seq(5, 35, 5)),
  YOB = NULL,
  observationYear = NULL,
  ...
)
}
\arguments{
\item{axes}{List of paramters spanning the dimensions of the grid.}

\item{YOB}{optional year of bith. If missing, the \code{observationYear} and the contract's age}

\item{observationYear}{The observation year, for which the grid shall be calculated. If given, the YOB is calculated from it, otherwise the contract's YOB is used}

\item{...}{Additional parameters to be passed to \link{InsuranceContract$new}}
}
\description{
Create a multi-dimensional grid of InsuranceContract objects, where the axes
ranges are given with the axes argument.
This function will return the full InsuranceContract objects, so apply can
later be used to extract premiums, reserves and other values to display in
a grid.
}
