% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractGrid.R
\name{contractGridPremium}
\alias{contractGridPremium}
\title{Create a multi-dimensional grid of premiums for insurance contracts, where the axes
ranges are given with the axes argument.
This function will return the full InsuranceContract objects, so apply can
later be used to extract premiums, reserves and other values to display in
a grid.}
\usage{
contractGridPremium(contractGrid = NULL, premium = "written", ...)
}
\description{
Create a multi-dimensional grid of premiums for insurance contracts, where the axes
ranges are given with the axes argument.
This function will return the full InsuranceContract objects, so apply can
later be used to extract premiums, reserves and other values to display in
a grid.
}
