% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\name{costsDisplayTable}
\alias{costsDisplayTable}
\title{Helper function to display all cost definitions in a concise table}
\usage{
costsDisplayTable(costs)
}
\arguments{
\item{costs}{The cost structure to be displayed in a concise table style.}
}
\description{
Returns a data.frame with columns
}
