% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{fillFields}
\alias{fillFields}
\title{Overwrite all existing fields in the first argument with
values given in valuelist. Members of valuelist that are not yet in
fields are ignored. This allows a huge valuelist to be used to fill
fields in multiple lists with given structure.}
\usage{
fillFields(fields, valuelist)
}
\arguments{
\item{fields}{existing list}

\item{valuelist}{list of fields to replace in \code{fields}. Only keys that exist in \code{fields} are overwritten, no new fields are added to \code{fields}}
}
\description{
Overwrite all existing fields in the first argument with
values given in valuelist. Members of valuelist that are not yet in
fields are ignored. This allows a huge valuelist to be used to fill
fields in multiple lists with given structure.
}
