% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{head0}
\alias{head0}
\title{Set all entries of the given vector to 0 up until index 'start'}
\usage{
head0(v, start = 0)
}
\description{
Set all entries of the given vector to 0 up until index 'start'
}
