% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceParameters.R
\name{initializeCosts}
\alias{initializeCosts}
\title{Initialize a cost matrix with dimensions: [CostType, Basis, Period], with:
    CostType: alpha, Zillmer, beta, gamma, gamma_nopremiums
    Basis:    SumInsured, SumPremiums, GrossPremium
    Period:   once, PremiumPeriod, PremiumFree, PolicyPeriod
TODO: gamma an Erlebensleistungen?}
\usage{
initializeCosts()
}
\description{
Initialize a cost matrix with dimensions: [CostType, Basis, Period], with:
    CostType: alpha, Zillmer, beta, gamma, gamma_nopremiums
    Basis:    SumInsured, SumPremiums, GrossPremium
    Period:   once, PremiumPeriod, PremiumFree, PolicyPeriod
TODO: gamma an Erlebensleistungen?
}

