% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceContract.R
\docType{data}
\name{InsuranceContract}
\alias{InsuranceContract}
\title{Base Class for Insurance Contracts}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InsuranceContract
}
\description{
R6 class that models a complete, general insurance contract.
The corresponding tariff and the profit participation scheme, as well as
all other relevant contract parameters (if not defined by the tariff or
explicitly overridden by the contract) can be given in the constructor.
}
\details{
Immediately upon construction, all premiums, reserves and cash flows for the
whole contract period are calculated.
}
\keyword{datasets}

