% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsuranceTarif.R
\name{InsuranceTarif}
\alias{InsuranceTarif}
\title{Base class for Insurance Tarifs, providing calculation functions to the contract}
\description{
This is a base class for holding contract-independent values and
providing methods to calculate cash flows, premiums, etc. Objects of this
class do NOT contain contract-specific values like age, death probabilities,
premiums, reserves, etc. Rather, they are the calculation kernels that will
be called by the \code{\link{InsuranceContract}} objects to make the actual,
tariff-specific calculations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InsuranceTarif$new()}}
\item \href{#method-createModification}{\code{InsuranceTarif$createModification()}}
\item \href{#method-getParameters}{\code{InsuranceTarif$getParameters()}}
\item \href{#method-getInternalValues}{\code{InsuranceTarif$getInternalValues()}}
\item \href{#method-getAges}{\code{InsuranceTarif$getAges()}}
\item \href{#method-getTransitionProbabilities}{\code{InsuranceTarif$getTransitionProbabilities()}}
\item \href{#method-getCostValues}{\code{InsuranceTarif$getCostValues()}}
\item \href{#method-getPremiumCF}{\code{InsuranceTarif$getPremiumCF()}}
\item \href{#method-getAnnuityCF}{\code{InsuranceTarif$getAnnuityCF()}}
\item \href{#method-getDeathCF}{\code{InsuranceTarif$getDeathCF()}}
\item \href{#method-getBasicCashFlows}{\code{InsuranceTarif$getBasicCashFlows()}}
\item \href{#method-getCashFlows}{\code{InsuranceTarif$getCashFlows()}}
\item \href{#method-getCashFlowsCosts}{\code{InsuranceTarif$getCashFlowsCosts()}}
\item \href{#method-presentValueCashFlows}{\code{InsuranceTarif$presentValueCashFlows()}}
\item \href{#method-presentValueCashFlowsCosts}{\code{InsuranceTarif$presentValueCashFlowsCosts()}}
\item \href{#method-costValuesAsMatrix}{\code{InsuranceTarif$costValuesAsMatrix()}}
\item \href{#method-getAbsCashFlows}{\code{InsuranceTarif$getAbsCashFlows()}}
\item \href{#method-getAbsPresentValues}{\code{InsuranceTarif$getAbsPresentValues()}}
\item \href{#method-presentValueBenefits}{\code{InsuranceTarif$presentValueBenefits()}}
\item \href{#method-getPremiumCoefficients}{\code{InsuranceTarif$getPremiumCoefficients()}}
\item \href{#method-premiumCalculation}{\code{InsuranceTarif$premiumCalculation()}}
\item \href{#method-reserveCalculation}{\code{InsuranceTarif$reserveCalculation()}}
\item \href{#method-getBalanceSheetReserveFactor}{\code{InsuranceTarif$getBalanceSheetReserveFactor()}}
\item \href{#method-reserveCalculationBalanceSheet}{\code{InsuranceTarif$reserveCalculationBalanceSheet()}}
\item \href{#method-calculateProfitParticipation}{\code{InsuranceTarif$calculateProfitParticipation()}}
\item \href{#method-reservesAfterProfit}{\code{InsuranceTarif$reservesAfterProfit()}}
\item \href{#method-getBasicDataTimeseries}{\code{InsuranceTarif$getBasicDataTimeseries()}}
\item \href{#method-premiumDecomposition}{\code{InsuranceTarif$premiumDecomposition()}}
\item \href{#method-calculateFutureSums}{\code{InsuranceTarif$calculateFutureSums()}}
\item \href{#method-calculatePresentValues}{\code{InsuranceTarif$calculatePresentValues()}}
\item \href{#method-clone}{\code{InsuranceTarif$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$new(
  name = NULL,
  type = "wholelife",
  tarif = "Generic Tarif",
  desc = "Description of tarif",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createModification"></a>}}
\if{latex}{\out{\hypertarget{method-createModification}{}}}
\subsection{Method \code{createModification()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$createModification(
  name = NULL,
  tarif = NULL,
  desc = NULL,
  tariffType = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getParameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInternalValues"></a>}}
\if{latex}{\out{\hypertarget{method-getInternalValues}{}}}
\subsection{Method \code{getInternalValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getInternalValues(params, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAges"></a>}}
\if{latex}{\out{\hypertarget{method-getAges}{}}}
\subsection{Method \code{getAges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getAges(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTransitionProbabilities"></a>}}
\if{latex}{\out{\hypertarget{method-getTransitionProbabilities}{}}}
\subsection{Method \code{getTransitionProbabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getTransitionProbabilities(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCostValues"></a>}}
\if{latex}{\out{\hypertarget{method-getCostValues}{}}}
\subsection{Method \code{getCostValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getCostValues(costs, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPremiumCF"></a>}}
\if{latex}{\out{\hypertarget{method-getPremiumCF}{}}}
\subsection{Method \code{getPremiumCF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getPremiumCF(len, params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAnnuityCF"></a>}}
\if{latex}{\out{\hypertarget{method-getAnnuityCF}{}}}
\subsection{Method \code{getAnnuityCF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getAnnuityCF(len, params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeathCF"></a>}}
\if{latex}{\out{\hypertarget{method-getDeathCF}{}}}
\subsection{Method \code{getDeathCF()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getDeathCF(len, params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBasicCashFlows"></a>}}
\if{latex}{\out{\hypertarget{method-getBasicCashFlows}{}}}
\subsection{Method \code{getBasicCashFlows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getBasicCashFlows(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCashFlows"></a>}}
\if{latex}{\out{\hypertarget{method-getCashFlows}{}}}
\subsection{Method \code{getCashFlows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getCashFlows(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCashFlowsCosts"></a>}}
\if{latex}{\out{\hypertarget{method-getCashFlowsCosts}{}}}
\subsection{Method \code{getCashFlowsCosts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getCashFlowsCosts(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-presentValueCashFlows"></a>}}
\if{latex}{\out{\hypertarget{method-presentValueCashFlows}{}}}
\subsection{Method \code{presentValueCashFlows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$presentValueCashFlows(cashFlows, params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-presentValueCashFlowsCosts"></a>}}
\if{latex}{\out{\hypertarget{method-presentValueCashFlowsCosts}{}}}
\subsection{Method \code{presentValueCashFlowsCosts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$presentValueCashFlowsCosts(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-costValuesAsMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-costValuesAsMatrix}{}}}
\subsection{Method \code{costValuesAsMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$costValuesAsMatrix(costValues)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAbsCashFlows"></a>}}
\if{latex}{\out{\hypertarget{method-getAbsCashFlows}{}}}
\subsection{Method \code{getAbsCashFlows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getAbsCashFlows(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAbsPresentValues"></a>}}
\if{latex}{\out{\hypertarget{method-getAbsPresentValues}{}}}
\subsection{Method \code{getAbsPresentValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getAbsPresentValues(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-presentValueBenefits"></a>}}
\if{latex}{\out{\hypertarget{method-presentValueBenefits}{}}}
\subsection{Method \code{presentValueBenefits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$presentValueBenefits(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPremiumCoefficients"></a>}}
\if{latex}{\out{\hypertarget{method-getPremiumCoefficients}{}}}
\subsection{Method \code{getPremiumCoefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getPremiumCoefficients(
  type = "gross",
  coeffBenefits,
  coeffCosts,
  premiums,
  params,
  values,
  premiumCalculationTime = values$int$premiumCalculationTime
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-premiumCalculation"></a>}}
\if{latex}{\out{\hypertarget{method-premiumCalculation}{}}}
\subsection{Method \code{premiumCalculation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$premiumCalculation(
  params,
  values,
  premiumCalculationTime = values$int$premiumCalculationTime
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reserveCalculation"></a>}}
\if{latex}{\out{\hypertarget{method-reserveCalculation}{}}}
\subsection{Method \code{reserveCalculation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$reserveCalculation(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBalanceSheetReserveFactor"></a>}}
\if{latex}{\out{\hypertarget{method-getBalanceSheetReserveFactor}{}}}
\subsection{Method \code{getBalanceSheetReserveFactor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getBalanceSheetReserveFactor(params, years = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reserveCalculationBalanceSheet"></a>}}
\if{latex}{\out{\hypertarget{method-reserveCalculationBalanceSheet}{}}}
\subsection{Method \code{reserveCalculationBalanceSheet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$reserveCalculationBalanceSheet(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateProfitParticipation"></a>}}
\if{latex}{\out{\hypertarget{method-calculateProfitParticipation}{}}}
\subsection{Method \code{calculateProfitParticipation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$calculateProfitParticipation(params, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reservesAfterProfit"></a>}}
\if{latex}{\out{\hypertarget{method-reservesAfterProfit}{}}}
\subsection{Method \code{reservesAfterProfit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$reservesAfterProfit(profitScenario, params, values, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBasicDataTimeseries"></a>}}
\if{latex}{\out{\hypertarget{method-getBasicDataTimeseries}{}}}
\subsection{Method \code{getBasicDataTimeseries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$getBasicDataTimeseries(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-premiumDecomposition"></a>}}
\if{latex}{\out{\hypertarget{method-premiumDecomposition}{}}}
\subsection{Method \code{premiumDecomposition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$premiumDecomposition(params, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateFutureSums"></a>}}
\if{latex}{\out{\hypertarget{method-calculateFutureSums}{}}}
\subsection{Method \code{calculateFutureSums()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$calculateFutureSums(values, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculatePresentValues"></a>}}
\if{latex}{\out{\hypertarget{method-calculatePresentValues}{}}}
\subsection{Method \code{calculatePresentValues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$calculatePresentValues(values, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InsuranceTarif$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
