% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{isSinglePremiumContract}
\alias{isSinglePremiumContract}
\title{Determine whether a contract (given all parameters) is a single-premium contract or with regular premiums}
\usage{
isSinglePremiumContract(params, values)
}
\arguments{
\item{params}{The parameters of the contract.}

\item{values}{Unused by default (already calculated values of the contract)}
}
\description{
Single premium contracts are identified by the parameter \code{premiumPeriod = 1}.
}
