% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractGrid.R
\name{makeContractGridDimname}
\alias{makeContractGridDimname}
\title{Generate a dimension label for the object passed as \code{value}, to be used in \link{contractGrid}}
\usage{
makeContractGridDimname(value)
}
\arguments{
\item{value}{the value along the axis, for which a name should be generated}
}
\description{
Generate a dimension label for the object passed as \code{value}, to be used in \link{contractGrid}
}
