% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractGrid.R
\name{makeContractGridDimnames}
\alias{makeContractGridDimnames}
\title{Generate proper dimnames for all entries of the axes of a \code{\link[=contractGrid]{contractGrid()}}}
\usage{
makeContractGridDimnames(axes)
}
\arguments{
\item{axes}{the axes with all names, for which a name should be generated}
}
\description{
Generate proper dimnames for all entries of the axes of a \code{\link[=contractGrid]{contractGrid()}}
}
\section{Functions}{
\itemize{
\item \code{makeContractGridDimnames}: Generate proper dimnames for all entries of the axes of a \code{\link[=contractGrid]{contractGrid()}}
}}

