% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{padLast}
\alias{padLast}
\title{Pad the vector \code{v} to length \code{l} by repeating the last entry of the
vector.}
\usage{
padLast(v, l, start = 0)
}
\arguments{
\item{v}{the vector to pad by repeating the last element}

\item{l}{the desired (resulting) length of the vector}

\item{start}{the first \code{start} values are always set to 0 (default is 0),
the vector \code{v} starts only after these leading zeroes. The number of
leading zeroes counts towards the desired length}
}
\description{
This function callc \code{\link[=pad0]{pad0()}} with the last element of the vector as padding value
}
\examples{
padLast(1:5, 7) # 5 is repeated twice
padLast(1:5, 3) # no padding needed

}
