% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfitParticipation_Functions.R, R/ProfitParticipation.R
\docType{data}
\name{profPart.base.PreviousZillmerReserve}
\alias{ProfitParticipation}
\alias{profPart.base.PreviousZillmerReserve}
\alias{profPart.base.ZillmerReserveT2}
\alias{profPart.base.ZillmerRiskPremium}
\alias{profPart.base.meanContractualReserve}
\alias{profPart.base.sumInsured}
\alias{profPart.calculate.RateOnBase}
\alias{profPart.rate.totalInterest}
\alias{profPart.rate.totalInterest2}
\title{Base Class for Profit Participation Schemes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
profPart.base.PreviousZillmerReserve(params, values)

profPart.base.ZillmerReserveT2(params, values)

profPart.base.meanContractualReserve(params, values)

profPart.base.ZillmerRiskPremium(params, values)

profPart.base.sumInsured(params, values)

profPart.rate.totalInterest(rates, params, values)

profPart.rate.totalInterest2(rates, params, values)

profPart.calculate.RateOnBase(base, rate, waiting, params, values)

ProfitParticipation
}
\description{
base class for Profit Participation schemes  (holding contract-independent values and
providing methods to calculate the profit participation values from the given
reserves).
}
\section{Functions}{
\itemize{
\item \code{profPart.base.PreviousZillmerReserve}: Basis for profit: Previous Zillmer reserve (no administration cost reserve)

\item \code{profPart.base.ZillmerReserveT2}: Basis for profit: Zillmer reserve (no administration cost reserve) at time t-2

\item \code{profPart.base.meanContractualReserve}: Basis for profit: Contractual reserve (including administration costs) averaged over t and t-1

\item \code{profPart.base.ZillmerRiskPremium}: Basis for risk/mortality profit: Zillmer Risk Premium of the past year

\item \code{profPart.base.sumInsured}: Basis for expense/sum profit: sum insured

\item \code{profPart.rate.totalInterest}: Rate for interest on past profits: total interest rate

\item \code{profPart.rate.totalInterest2}: Rate for interest on past profits: second total interest rate

\item \code{profPart.calculate.RateOnBase}: Calculate profit by a simple rate applied on the basis (with an optional waiting vector of values 0 or 1)
}}
\keyword{datasets}

