% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfitParticipation.R
\docType{data}
\name{ProfitParticipation}
\alias{ProfitParticipation}
\title{Base Class for Profit Participation Schemes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProfitParticipation
}
\description{
base class for Profit Participation schemes  (holding contract-independent values and
providing methods to calculate the profit participation values from the given
reserves).
}
\keyword{datasets}

