% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\docType{class}
\name{SexSingleEnum-class}
\alias{SexSingleEnum-class}
\alias{SexEnum}
\title{Enum to describe possible sexes in an insurance contract or tariff.}
\description{
Enum to describe possible sexes in an insurance contract or tariff.
}
\details{
Currently, the only possible values are:
\itemize{
\item "unisex"
\item "male"
\item "female"
}
}
